lwext4_rust

Module bindings

Structs

_IO_FILE

__IncompleteArrayField

__va_list_tag

ext4_bcache

@brief Block cache descriptor

ext4_bcache_ext4_buf_dirty

@brief A singly-linked list holding dirty buffers

ext4_bcache_ext4_buf_lba

@brief A tree holding all bufs

ext4_bcache_ext4_buf_lru

@brief A tree holding unreferenced bufs

ext4_bgroup

ext4_block

@brief Single block descriptor

ext4_block_group_ref

ext4_blockdev

@brief Definition of the simple block device.

ext4_blockdev_iface

ext4_buf

@brief Single block descriptor

ext4_buf__bindgen_ty_1

ext4_buf__bindgen_ty_2

ext4_buf__bindgen_ty_3

ext4_dir

@brief Directory descriptor.

ext4_dir_en

Linked list directory entry structure

ext4_dir_entry_tail

ext4_dir_idx_climit

ext4_dir_idx_dot_en

ext4_dir_idx_entry

ext4_dir_idx_node

ext4_dir_idx_rinfo

ext4_dir_idx_root

ext4_dir_idx_tail

ext4_direntry

@brief Directory entry descriptor.

ext4_fake_dir_entry

ext4_file

@brief File descriptor.

ext4_fs

ext4_inode

ext4_inode__bindgen_ty_1__bindgen_ty_1

ext4_inode__bindgen_ty_1__bindgen_ty_2

ext4_inode_ref

ext4_io_stats

ext4_lock

@brief OS dependent lock interface.

ext4_mbr_bdevs

@brief Master boot record block devices descriptor

ext4_mbr_parts

@brief Master boot record partitions

ext4_mkfs_info

ext4_mount_stats

@brief Some of the filesystem stats.

ext4_mountpoint

@brief Mount point handle.

ext4_sblock

imaxdiv_t

jbd_bhdr

jbd_block_tag

jbd_block_tag3

jbd_block_tail

jbd_commit_header

jbd_fs

jbd_journal

jbd_revoke_header

jbd_revoke_tail

jbd_sb

jbd_trans

max_align_t

Constants

BUFSIZ

CONFIG_BLOCK_DEV_CACHE_SIZE

CONFIG_BLOCK_DEV_ENABLE_STATS

CONFIG_DEBUG_ASSERT

CONFIG_DEBUG_PRINTF

CONFIG_EXT4_BLOCKDEVS_COUNT

CONFIG_EXT4_MAX_BLOCKDEV_NAME

CONFIG_EXT4_MAX_MP_NAME

CONFIG_EXT4_MOUNTPOINTS_COUNT

CONFIG_EXTENTS_ENABLE

CONFIG_EXT_FEATURE_SET_LVL

CONFIG_HAVE_OWN_ASSERT

CONFIG_HAVE_OWN_ERRNO

CONFIG_HAVE_OWN_OFLAGS

CONFIG_JOURNALING_ENABLE

CONFIG_MAX_TRUNCATE_SIZE

CONFIG_UNALIGNED_ACCESS

CONFIG_USE_DEFAULT_CONFIG

CONFIG_USE_USER_MALLOC

CONFIG_XATTR_ENABLE

DBG_ERROR

DBG_INFO

DBG_NONE

DBG_WARN

DEBUG_ALL

DEBUG_BALLOC

DEBUG_BCACHE

DEBUG_BITMAP

DEBUG_BLOCKDEV

DEBUG_BLOCK_GROUP

DEBUG_DIR

DEBUG_DIR_IDX

DEBUG_EXT4

DEBUG_EXTENT

DEBUG_FS

DEBUG_HASH

DEBUG_IALLOC

DEBUG_INODE

DEBUG_JBD

DEBUG_MBR

DEBUG_MKFS

DEBUG_NOPREFIX

DEBUG_SUPER

DEBUG_XATTR

E2BIG

EACCES

EEXIST

EFAULT

EFBIG

EINVAL

EIO

EISDIR

EMLINK

ENODATA

ENODEV

ENOENT

ENOMEM

ENOSPC

ENOTDIR

ENOTEMPTY

ENOTSUP

ENXIO

EOF

EOK

EPERM

ERANGE

EROFS

EXT2_HTREE_EOF

EXT2_HTREE_HALF_MD4

EXT2_HTREE_HALF_MD4_UNSIGNED

EXT2_HTREE_LEGACY

EXT2_HTREE_LEGACY_UNSIGNED

EXT2_HTREE_TEA

EXT2_HTREE_TEA_UNSIGNED

EXT2_SUPPORTED_FCOM

EXT2_SUPPORTED_FINCOM

EXT2_SUPPORTED_FRO_COM

EXT3_SUPPORTED_FCOM

EXT3_SUPPORTED_FINCOM

EXT3_SUPPORTED_FRO_COM

EXT4_BAD_INO

EXT4_BLOCK_GROUP_BLOCK_UNINIT

EXT4_BLOCK_GROUP_INODE_UNINIT

EXT4_BLOCK_GROUP_ITABLE_ZEROED

EXT4_BOOT_LOADER_INO

EXT4_CHECKSUM_CRC32C

EXT4_CRC32_INIT

EXT4_DE_BLKDEV

EXT4_DE_CHRDEV

EXT4_DE_DIR

EXT4_DE_FIFO

EXT4_DE_REG_FILE

EXT4_DE_SOCK

EXT4_DE_SYMLINK

EXT4_DE_UNKNOWN

EXT4_DIRECTORY_FILENAME_LEN

EXT4_DIRENTRY_DIR_CSUM

EXT4_ERR_BAD_DX_DIR

EXT4_FCOM_DIR_INDEX

EXT4_FCOM_DIR_PREALLOC

EXT4_FCOM_EXT_ATTR

EXT4_FCOM_HAS_JOURNAL

EXT4_FCOM_IMAGIC_INODES

EXT4_FCOM_RESIZE_INODE

EXT4_FINCOM_64BIT

EXT4_FINCOM_BG_USE_META_CSUM

EXT4_FINCOM_COMPRESSION

EXT4_FINCOM_DIRDATA

EXT4_FINCOM_EA_INODE

EXT4_FINCOM_EXTENTS

EXT4_FINCOM_FILETYPE

EXT4_FINCOM_FLEX_BG

EXT4_FINCOM_INLINE_DATA

EXT4_FINCOM_JOURNAL_DEV

EXT4_FINCOM_LARGEDIR

EXT4_FINCOM_META_BG

EXT4_FINCOM_MMP

EXT4_FINCOM_RECOVER

EXT4_FRO_COM_BIGALLOC

EXT4_FRO_COM_BTREE_DIR

EXT4_FRO_COM_DIR_NLINK

EXT4_FRO_COM_EXTRA_ISIZE

EXT4_FRO_COM_GDT_CSUM

EXT4_FRO_COM_HUGE_FILE

EXT4_FRO_COM_LARGE_FILE

EXT4_FRO_COM_METADATA_CSUM

EXT4_FRO_COM_QUOTA

EXT4_FRO_COM_SPARSE_SUPER

EXT4_GOOD_OLD_FIRST_INO

EXT4_GOOD_OLD_INODE_SIZE

EXT4_INODE_BLOCKS

EXT4_INODE_BLOCK_SIZE

EXT4_INODE_DIRECT_BLOCK_COUNT

EXT4_INODE_DOUBLE_INDIRECT_BLOCK

EXT4_INODE_FLAG_APPEND

EXT4_INODE_FLAG_COMPR

EXT4_INODE_FLAG_COMPRBLK

EXT4_INODE_FLAG_DIRSYNC

EXT4_INODE_FLAG_DIRTY

EXT4_INODE_FLAG_EA_INODE

EXT4_INODE_FLAG_ECOMPR

EXT4_INODE_FLAG_EOFBLOCKS

EXT4_INODE_FLAG_EXTENTS

EXT4_INODE_FLAG_HUGE_FILE

EXT4_INODE_FLAG_IMAGIC

EXT4_INODE_FLAG_IMMUTABLE

EXT4_INODE_FLAG_INDEX

EXT4_INODE_FLAG_JOURNAL_DATA

EXT4_INODE_FLAG_NOATIME

EXT4_INODE_FLAG_NOCOMPR

EXT4_INODE_FLAG_NODUMP

EXT4_INODE_FLAG_NOTAIL

EXT4_INODE_FLAG_RESERVED

EXT4_INODE_FLAG_SECRM

EXT4_INODE_FLAG_SYNC

EXT4_INODE_FLAG_TOPDIR

EXT4_INODE_FLAG_UNRM

EXT4_INODE_INDIRECT_BLOCK

EXT4_INODE_INDIRECT_BLOCK_COUNT

EXT4_INODE_MODE_BLOCKDEV

EXT4_INODE_MODE_CHARDEV

EXT4_INODE_MODE_DIRECTORY

EXT4_INODE_MODE_FIFO

EXT4_INODE_MODE_FILE

EXT4_INODE_MODE_SOCKET

EXT4_INODE_MODE_SOFTLINK

EXT4_INODE_MODE_TYPE_MASK

EXT4_INODE_ROOT_INDEX

EXT4_INODE_TRIPPLE_INDIRECT_BLOCK

EXT4_JOURNAL_INO

EXT4_LINK_MAX

EXT4_MAX_BLOCK_GROUP_DESCRIPTOR_SIZE

EXT4_MAX_BLOCK_SIZE

EXT4_MIN_BLOCK_GROUP_DESCRIPTOR_SIZE

EXT4_MIN_BLOCK_SIZE

EXT4_RESIZE_INO

EXT4_REV0_INODE_SIZE

EXT4_ROOT_INO

EXT4_SUPERBLOCK_ERRORS_CONTINUE

EXT4_SUPERBLOCK_ERRORS_PANIC

EXT4_SUPERBLOCK_ERRORS_RO

EXT4_SUPERBLOCK_FLAGS_SIGNED_HASH

EXT4_SUPERBLOCK_FLAGS_TEST_FILESYS

EXT4_SUPERBLOCK_FLAGS_UNSIGNED_HASH

EXT4_SUPERBLOCK_MAGIC

EXT4_SUPERBLOCK_OFFSET

EXT4_SUPERBLOCK_OS_HURD

EXT4_SUPERBLOCK_OS_LINUX

EXT4_SUPERBLOCK_SIZE

EXT4_SUPERBLOCK_STATE_ERROR_FS

EXT4_SUPERBLOCK_STATE_ORPHAN_FS

EXT4_SUPERBLOCK_STATE_VALID_FS

EXT4_SUPPORTED_FCOM

EXT4_SUPPORTED_FINCOM

EXT4_SUPPORTED_FRO_COM

EXT4_UNDEL_DIR_INO

EXT_FINCOM_IGNORED

FILENAME_MAX

FOPEN_MAX

F_SET_EXT2

F_SET_EXT3

F_SET_EXT4

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INT64_MAX

INT64_MIN

INTMAX_MAX

INTMAX_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST64_MAX

INT_FAST64_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST64_MAX

INT_LEAST64_MIN

JBD_COMMIT_BLOCK

JBD_CRC32C_CHKSUM

JBD_CRC32_CHKSUM

JBD_CRC32_CHKSUM_SIZE

JBD_DESCRIPTOR_BLOCK

JBD_FEATURE_COMPAT_CHECKSUM

JBD_FEATURE_INCOMPAT_64BIT

JBD_FEATURE_INCOMPAT_ASYNC_COMMIT

JBD_FEATURE_INCOMPAT_CSUM_V2

JBD_FEATURE_INCOMPAT_CSUM_V3

JBD_FEATURE_INCOMPAT_REVOKE

JBD_FLAG_DELETED

JBD_FLAG_ESCAPE

JBD_FLAG_LAST_TAG

JBD_FLAG_SAME_UUID

JBD_KNOWN_COMPAT_FEATURES

JBD_KNOWN_INCOMPAT_FEATURES

JBD_KNOWN_ROCOMPAT_FEATURES

JBD_MAGIC_NUMBER

JBD_MD5_CHKSUM

JBD_REVOKE_BLOCK

JBD_SHA1_CHKSUM

JBD_SUPERBLOCK

JBD_SUPERBLOCK_V2

JBD_USERS_MAX

JBD_USERS_SIZE

L_ctermid

L_cuserid

L_tmpnam

O_APPEND

O_CREAT

O_EXCL

O_RDONLY

O_RDWR

O_TRUNC

O_WRONLY

PRIX8

PRIX16

PRIX32

PRIX64

PRIXFAST8

PRIXFAST16

PRIXFAST32

PRIXFAST64

PRIXLEAST8

PRIXLEAST16

PRIXLEAST32

PRIXLEAST64

PRIXMAX

PRIXPTR

PRId8

PRId16

PRId32

PRId64

PRIdFAST8

PRIdFAST16

PRIdFAST32

PRIdFAST64

PRIdLEAST8

PRIdLEAST16

PRIdLEAST32

PRIdLEAST64

PRIdMAX

PRIdPTR

PRIi8

PRIi16

PRIi32

PRIi64

PRIiFAST8

PRIiFAST16

PRIiFAST32

PRIiFAST64

PRIiLEAST8

PRIiLEAST16

PRIiLEAST32

PRIiLEAST64

PRIiMAX

PRIiPTR

PRIo8

PRIo16

PRIo32

PRIo64

PRIoFAST8

PRIoFAST16

PRIoFAST32

PRIoFAST64

PRIoLEAST8

PRIoLEAST16

PRIoLEAST32

PRIoLEAST64

PRIoMAX

PRIoPTR

PRIu8

PRIu16

PRIu32

PRIu64

PRIuFAST8

PRIuFAST16

PRIuFAST32

PRIuFAST64

PRIuLEAST8

PRIuLEAST16

PRIuLEAST32

PRIuLEAST64

PRIuMAX

PRIuPTR

PRIx8

PRIx16

PRIx32

PRIx64

PRIxFAST8

PRIxFAST16

PRIxFAST32

PRIxFAST64

PRIxLEAST8

PRIxLEAST16

PRIxLEAST32

PRIxLEAST64

PRIxMAX

PRIxPTR

PTRDIFF_MAX

PTRDIFF_MIN

P_tmpdir

RB_BLACK

RB_INF

RB_NEGINF

RB_RED

SCNd8

SCNd16

SCNd32

SCNd64

SCNdFAST8

SCNdFAST16

SCNdFAST32

SCNdFAST64

SCNdLEAST8

SCNdLEAST16

SCNdLEAST32

SCNdLEAST64

SCNdMAX

SCNdPTR

SCNi8

SCNi16

SCNi32

SCNi64

SCNiFAST8

SCNiFAST16

SCNiFAST32

SCNiFAST64

SCNiLEAST8

SCNiLEAST16

SCNiLEAST32

SCNiLEAST64

SCNiMAX

SCNiPTR

SCNo8

SCNo16

SCNo32

SCNo64

SCNoFAST8

SCNoFAST16

SCNoFAST32

SCNoFAST64

SCNoLEAST8

SCNoLEAST16

SCNoLEAST32

SCNoLEAST64

SCNoMAX

SCNoPTR

SCNu8

SCNu16

SCNu32

SCNu64

SCNuFAST8

SCNuFAST16

SCNuFAST32

SCNuFAST64

SCNuLEAST8

SCNuLEAST16

SCNuLEAST32

SCNuLEAST64

SCNuMAX

SCNuPTR

SCNx8

SCNx16

SCNx32

SCNx64

SCNxFAST8

SCNxFAST16

SCNxFAST32

SCNxFAST64

SCNxLEAST8

SCNxLEAST16

SCNxLEAST32

SCNxLEAST64

SCNxMAX

SCNxPTR

SEEK_CUR

SEEK_END

SEEK_SET

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

SPLAY_INF

SPLAY_NEGINF

TMP_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINT64_MAX

UINTMAX_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_FAST64_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

UINT_LEAST64_MAX

UUID_SIZE

WINT_MAX

WINT_MIN

_BSD_SOURCE

_IOFBF

_IOLBF

_IONBF

_XOPEN_SOURCE

__BIG_ENDIAN

__BYTE_ORDER

__LITTLE_ENDIAN

__LONG_MAX

__PRI64

__PRIPTR

__USE_TIME_BITS64

__bool_true_false_are_defined

bcache_state_bits_BC_DIRTY

bcache_state_bits_BC_FLUSH

bcache_state_bits_BC_TMP

bcache_state_bits_BC_UPTODATE

false_

true_

Statics

stderrโ

stdinโ

stdoutโ

Functions

asprintfโ

clearerrโ

clearerr_unlockedโ

ctermidโ

cuseridโ

dprintfโ

ext4_atime_getโ

@brief Get file/directory/link access time.

ext4_atime_setโ

@brief Set file/directory/link access time.

ext4_bcache_allocโ

@brief Allocate block from block cache memory. Unreferenced block allocation is based on LRU (Last Recently Used) algorithm. @param bc block cache descriptor @param b block to alloc @param is_new block is new (needs to be read) @return standard error code

ext4_bcache_cleanupโ

@brief Do cleanup works on block cache. @param bc block cache descriptor.

ext4_bcache_drop_bufโ

@brief Drop unreferenced buffer from bcache. @param bc block cache descriptor @param buf buffer

ext4_bcache_find_getโ

@brief Find existing buffer from block cache memory. Unreferenced block allocation is based on LRU (Last Recently Used) algorithm. @param bc block cache descriptor @param b block to alloc @param lba logical block address @return block cache buffer

ext4_bcache_fini_dynamicโ

@brief Dynamic de-initialization of block cache. @param bc block cache descriptor @return standard error code

ext4_bcache_freeโ

@brief Free block from cache memory (decrement reference counter). @param bc block cache descriptor @param b block to free @return standard error code

ext4_bcache_init_dynamicโ

@brief Dynamic initialization of block cache. @param bc block cache descriptor @param cnt items count in block cache @param itemsize single item size (in bytes) @return standard error code

ext4_bcache_invalidate_bufโ

@brief Invalidate a buffer. @param bc block cache descriptor @param buf buffer

ext4_bcache_invalidate_lbaโ

@brief Invalidate a range of buffers. @param bc block cache descriptor @param from starting lba @param cnt block counts

ext4_bcache_is_fullโ

@brief Return a full status of block cache. @param bc block cache descriptor @return full status

ext4_block_bind_bcacheโ

@brief Binds a bcache to block device. @param bdev block device descriptor @param bc block cache descriptor @return standard error code

ext4_block_cache_flushโ

@brief Flush all dirty buffers to disk @param bdev block device descriptor @return standard error code

ext4_block_cache_write_backโ

@brief Enable/disable write back cache mode @param bdev block device descriptor @param on_off !0 - ENABLE 0 - DISABLE (all delayed cache buffers will be flushed) @return standard error code

ext4_block_finiโ

@brief Close block device @param bdev block device descriptor @return standard error code

ext4_block_flush_bufโ

@brief Flush data in given buffer to disk. @param bdev block device descriptor @param buf buffer @return standard error code

ext4_block_flush_lbaโ

@brief Flush data in buffer of given lba to disk, if that buffer exists in block cache. @param bdev block device descriptor @param lba logical block address @return standard error code

ext4_block_getโ

@brief Block get function (through cache). @param bdev block device descriptor @param b block descriptor @param lba logical block address @return standard error code

ext4_block_get_noreadโ

@brief Block get function (through cache, donโt read). @param bdev block device descriptor @param b block descriptor @param lba logical block address @return standard error code

ext4_block_initโ

@brief Block device initialization. @param bdev block device descriptor @return standard error code

ext4_block_readbytesโ

@brief Read freom block device (by direct address). @param bdev block device descriptor @param off byte offset in block device @param buf input buffer @param len length of the write buffer @return standard error code

ext4_block_setโ

@brief Block set procedure (through cache). @param bdev block device descriptor @param b block descriptor @return standard error code

ext4_block_set_lb_sizeโ

@brief Set logical block size in block device. @param bdev block device descriptor @param lb_bsize logical block size (in bytes)

ext4_block_writebytesโ

@brief Write to block device (by direct address). @param bdev block device descriptor @param off byte offset in block device @param buf input buffer @param len length of the write buffer @return standard error code

ext4_blocks_get_directโ

@brief Block read procedure (without cache) @param bdev block device descriptor @param buf output buffer @param lba logical block address @return standard error code

ext4_blocks_set_directโ

@brief Block write procedure (without cache) @param bdev block device descriptor @param buf output buffer @param lba logical block address @return standard error code

ext4_buf_lowest_lruโ

@brief Get a buffer with the lowest LRU counter in bcache. @param bc block cache descriptor @return buffer with the lowest LRU counter

ext4_cache_flushโ

@brief Force cache flush.

ext4_cache_write_backโ

@brief Enable/disable write back cache mode. @warning Default model of cache is write trough. It means that when You do:

ext4_ctime_getโ

@brief Get file/directory/link change time.

ext4_ctime_setโ

@brief Set file/directory/link change time.

ext4_device_registerโ

@brief Register block device.

ext4_device_unregisterโ

@brief Un-register block device.

ext4_device_unregister_allโ

@brief Un-register all block devices.

ext4_dir_closeโ

@brief Directory close.

ext4_dir_entry_nextโ

@brief Return next directory entry.

ext4_dir_entry_rewindโ

@brief Rewine directory entry offset.

ext4_dir_mkโ

@brief Create new directory.

ext4_dir_mvโ

@brief Rename/move directory.

ext4_dir_openโ

@brief Directory open.

ext4_dir_rmโ

@brief Recursive directory remove.

ext4_dmask_clrโ

@brief Global mask debug clear. @param m new debug mask.

ext4_dmask_getโ

@brief Global debug mask get. @return debug mask

ext4_dmask_setโ

@brief Global mask debug set. @param m new debug mask.

ext4_fcloseโ

@brief File close function.

ext4_flinkโ

@brief Create a hardlink for a file.

ext4_fopenโ

@brief File open function.

ext4_fopen2โ

@brief Alternate file open function.

ext4_freadโ

@brief Read data from file.

ext4_fremoveโ

@brief Remove file by path.

ext4_frenameโ

@brief Rename file. @param path Source. @param new_path Destination. @return Standard error code.

ext4_fs_alloc_inodeโ

@brief Allocate new i-node in the filesystem. @param fs Filesystem to allocated i-node on @param inode_ref Output pointer to return reference to allocated i-node @param filetype File type of newly created i-node @return Error code

ext4_fs_append_inode_dblkโ

@brief Append following logical block to the i-node. @param inode_ref I-node to append block to @param fblock Output physical block address of newly allocated block @param iblock Output logical number of newly allocated block @return Error code

ext4_fs_check_featuresโ

@brief Check filesystemโs features, if supported by this driver Function can return EOK and set read_only flag. It meanโs that there are some not-supported features, that can cause problems during some write operations. @param fs Filesystem to be checked @param read_only Flag if filesystem should be mounted only for reading @return Error code

ext4_fs_correspond_inode_modeโ

@brief Convert filetype to inode mode. @param filetype File type @return inode mode

ext4_fs_finiโ

@brief Destroy filesystem instance (used by unmount operation). @param fs Filesystem to be destroyed @return Error code

ext4_fs_free_inodeโ

@brief Release i-node and mark it as free. @param inode_ref I-node to be released @return Error code

ext4_fs_get_block_group_refโ

@brief Get reference to block group specified by index. @param fs Filesystem to find block group on @param bgid Index of block group to load @param ref Output pointer for reference @return Error code

ext4_fs_get_inode_dblk_idxโ

@brief Get physical block address by logical index of the block. @param inode_ref I-node to read block address from @param iblock Logical index of block @param fblock Output pointer for return physical block address @param support_unwritten Indicate whether unwritten block range is supported under the current context @return Error code

ext4_fs_get_inode_refโ

@brief Get reference to i-node specified by index. @param fs Filesystem to find i-node on @param index Index of i-node to load @param ref Output pointer for reference @return Error code

ext4_fs_indirect_find_goalโ

@brief Compute โgoalโ for allocation algorithm (For blockmap). @param inode_ref Reference to inode, to allocate block for @return error code

ext4_fs_initโ

@brief Initialize filesystem and read all needed data. @param fs Filesystem instance to be initialized @param bdev Identifier if device with the filesystem @param read_only Mark the filesystem as read-only. @return Error code

ext4_fs_init_inode_dblk_idxโ

@brief Initialize a part of unwritten range of the inode. @param inode_ref I-node to proceed on. @param iblock Logical index of block @param fblock Output pointer for return physical block address @return Error code

ext4_fs_inode_blocks_initโ

@brief Reset blocks field of i-node. @param fs Filesystem to reset blocks field of i-inode on @param inode_ref ref Pointer for inode to be operated on

ext4_fs_inode_links_count_decโ

@brief Decrement inode link count. @param inode_ref none handle

ext4_fs_inode_links_count_incโ

@brief Increment inode link count. @param inode_ref none handle

ext4_fs_inode_to_goal_blockโ

@brief Compute โgoalโ for inode index @param inode_ref Reference to inode, to allocate block for @return goal

ext4_fs_put_block_group_refโ

@brief Put reference to block group. @param ref Pointer for reference to be put back @return Error code

ext4_fs_put_inode_refโ

@brief Put reference to i-node. @param ref Pointer for reference to be put back @return Error code

ext4_fs_truncate_inodeโ

@brief Truncate i-node data blocks. @param inode_ref I-node to be truncated @param new_size New size of inode (must be < current size) @return Error code

ext4_fseekโ

@brief File seek operation.

ext4_fsizeโ

@brief Get file size.

ext4_fsymlinkโ

@brief Create symbolic link.

ext4_ftellโ

@brief Get file position.

ext4_ftruncateโ

@brief File truncate function.

ext4_fwriteโ

@brief Write data to file.

ext4_get_sblockโ

@brief Acquire the filesystem superblock pointer of a mp.

ext4_getxattrโ

@brief Get extended attribute.

ext4_inode_existโ

@brief Check if inode exists.

ext4_journal_startโ

@brief Starts journaling. Journaling start/stop functions are transparent and might be used on filesystems without journaling support. @warning Usage: ext4_mount(โsda1โ, โ/โ); ext4_journal_start(โ/โ);

ext4_journal_stopโ

@brief Stops journaling. Journaling start/stop functions are transparent and might be used on filesystems without journaling support.

ext4_listxattrโ

@brief List extended attributes.

ext4_mbr_scanโ

ext4_mbr_writeโ

ext4_mkfsโ

ext4_mkfs_read_infoโ

ext4_mknodโ

@brief Create special file. @param path Path to new special file. @param filetype Filetype of the new special file. (that must not be regular file, directory, or unknown type) @param dev If filetype is char device or block device, the device number will become the payload in the inode. @return Standard error code.

ext4_mode_getโ

@brief Get file/directory/link mode bits.

ext4_mode_setโ

@brief Change file/directory/link mode bits.

ext4_mountโ

@brief Mount a block device with EXT4 partition to the mount point.

ext4_mount_point_statsโ

@brief Get file mount point stats.

ext4_mount_setup_locksโ

@brief Setup OS lock routines.

ext4_mtime_getโ

@brief Get file/directory/link modify time.

ext4_mtime_setโ

@brief Set file/directory/link modify time.

ext4_owner_getโ

@brief Get file/directory/link owner and group.

ext4_owner_setโ

@brief Change file owner and group.

ext4_raw_inode_fillโ

@brief Get inode of file/directory/link.

ext4_readlinkโ

@brief Read symbolic link payload.

ext4_recoverโ

@brief Journal recovery. @warning Must be called after @ref ext4_mount.

ext4_removexattrโ

@brief Remove extended attribute.

ext4_setxattrโ

@brief Set extended attribute.

ext4_umountโ

@brief Umount operation.

fcloseโ

fdopenโ

feofโ

feof_unlockedโ

ferrorโ

ferror_unlockedโ

fflushโ

fflush_unlockedโ

fgetcโ

fgetc_unlockedโ

fgetlnโ

fgetposโ

fgetsโ

filenoโ

fileno_unlockedโ

flockfileโ

fmemopenโ

fopenโ

fprintfโ

fputcโ

fputc_unlockedโ

fputsโ

freadโ

fread_unlockedโ

freopenโ

fscanfโ

fseekโ

fseekoโ

fsetposโ

ftellโ

ftelloโ

ftrylockfileโ

funlockfileโ

fwriteโ

fwrite_unlockedโ

getcโ

getc_unlockedโ

getcharโ

getchar_unlockedโ

getdelimโ

getlineโ

getwโ

imaxabsโ

imaxdivโ

io_timings_clearโ

io_timings_getโ

open_memstreamโ

pcloseโ

perrorโ

popenโ

printfโ

putcโ

putc_unlockedโ

putcharโ

putchar_unlockedโ

putsโ

putwโ

removeโ

renameโ

renameatโ

rewindโ

scanfโ

setbufโ

setbufferโ

setlinebufโ

setvbufโ

snprintfโ

sprintfโ

sscanfโ

strtoimaxโ

strtoumaxโ

tempnamโ

test_lwext4_block_statsโ

test_lwext4_cleanupโ

test_lwext4_dir_lsโ

test_lwext4_dir_testโ

test_lwext4_file_testโ

test_lwext4_mountโ

test_lwext4_mp_statsโ

test_lwext4_umountโ

tim_get_msโ

tim_get_usโ

tim_wait_msโ

tmpfileโ

tmpnamโ

ungetcโ

vasprintfโ

vdprintfโ

vfprintfโ

vfscanfโ

vprintfโ

vscanfโ

vsnprintfโ

vsprintfโ

vsscanfโ

wcstoimaxโ

wcstoumaxโ

Type Aliases

FILE

__builtin_va_list

__isoc_va_list

_bindgen_ty_1

@brief Directory entry types.

bcache_state_bits

@brief buffer state bits

ext4_fsblk_t

ext4_lblk_t

fpos_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

off_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

va_list

wchar_t

Unions

_G_fpos64_t

ext4_dir_en_internal

ext4_inode__bindgen_ty_1